detector: Rename it to protocoldetector
