+++ src/sensor/protocoldetector.h
